backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
requests==2.32.3
pyyaml==6.0.2